Actually set type tag in the setters
Summary:
As title. Note that emplace doesn't update the value if the key already exists (https://en.cppreference.com/w/cpp/container/unordered_map/emplace).

Not sure if this was causing any type of issue, but I checked that the key doesn't exist in the map before setting (always_assert(map.count(key) == 0)) and it failed.

I do expect that we want to update the value, in case it already existed. cc Wei. Thanks

Reviewed By: wsanville

Differential Revision: D13913796

fbshipit-source-id: dcaae067d13f7dadfe69f7b68da25f8d82bf8595