Converting ClassObjectSource back to ConstantAbstractDomain
Summary:
We are going back with a simpler design for tracking `ClassObjectSource`.
The lattice might be an overkill since most of the time we just need to distinguish between a result of a reflecting or a non-reflecting operation.
Whenever the information is UNKNOWN, the corresponding `ClassObjectSource` will simply be `boost::none`. Which requires additional handling.

Reviewed By: int3

Differential Revision: D13962252

fbshipit-source-id: 7b2fa7c15c56f9be7fa52330150abae7b3aa2daf