dcev2: Fix logic around removing <init>s
Summary:
Previously, we avoided removing `<init>`s if the object was ever read or written to, but neglected to do this if the object-bearing register was referenced by a non-move opcode. Concretely, given the following code:

```
new-instance v0 LFoo;
invoke-direct {v0} LFoo;.<init>();
if-eqz v0
// no more references to v0
```

We would remove the `invoke-direct` call but not the `new-instance` or `if-eqz` ones. This was a verification error -- `if-eqz` would be left referencing an uninitialized object. This diff fixes things so that we don't remove the invoke to the `<init>` (unless some other pass elides the `if-eqz` check first).

Reviewed By: justinjhendrick

Differential Revision: D13915473

fbshipit-source-id: 0ceddefb19ff64b74c1e91917a52d3d97e5655b9