Instr test for ClassObjectSource join
Summary:
Adding the test for the join operation between two `ClassObjectSource`s as an instr test.
This way we can feed more complicated bytecode compiled from Java source file into the analysis.

Since the join of a `REFLECTION` and a `NON_REFLECTION` is simply top, we don't have any *reflection sites* collected from the *join* block. The output is as expected.

Reviewed By: arnaudvenet

Differential Revision: D13945958

fbshipit-source-id: dd97719da3c9edfbf73b3af981dfb3769eb97a86