Promote D14017157 to redex-staging, redex-stable
Summary:
As per D14017157 (which landed in redex-unstable):

The method reordering optimization currently takes methods that occur in ~95%+ of beta user startups and moves them to the top of the first secondary dex. This means that when new experiments are launched which are not broadly rolled out, any new methods introduced in said experiments will usually not be at the top of the first secondary dex, and may hence be more expensive than other more favorable methods during startup.

This change introduces a manually overridable list of substrings that can match method names. This allows certain methods to be manually brought to the top of the first secondary dex irrespective of the method reordering file.

It initially contains `EagerInlineComposerSectionAdapter`, to try to rule out method reordering bias as a possible reason for the inconclusive performance result of [this](https://fb.workplace.com/groups/691614554220496/permalink/2273408272707775/?comment_id=2273430119372257&reply_comment_id=2273698466012089&comment_tracking=%7B%22tn%22%3A%22R%22%7D) experiment.

Reviewed By: minjang

Differential Revision: D14179851

fbshipit-source-id: f891083e576a4f4a63f533ccf7ab1ce19b360c2d