Handle classes with no package name in ClassOptData
Summary:
We seem to have some autogenerated classes that are being emitted without a package name: https://fburl.com/biggrep/vliko4u2

This doesn't seem to be a problem for our shipping apps, but at one point during a pass list bisect I disabled some dead code elimination, and the inliner choked while logging a decision via OptData due to this assert.

It seems only for logging purposes, so let it carry on with the empty string for this case.

Reviewed By: justinjhendrick

Differential Revision: D14238053

fbshipit-source-id: 644424ab9b667e4a398c517285944a5e55bd4d5b