Fix OptimizeEnums insertion of may_throw insns
Summary:
Fix an issue introduced in the previous diff where OptimizeEnums would run in an infinite loop. OptimizeEnums was adding an instruction that could throw while iterating through the CFG. This causes an infinite loop if the instruction is inserted inside a try region because the CFG needs to create a new block and move code around to remain in a valid state.

The fix is to defer the addition of instructions that may throw until after we're done with the analysis.

Also, add make the cfg insertion methods return a boolean denoting whether or not iterators were invalidated.

Reviewed By: Feng23

Differential Revision: D14248218

fbshipit-source-id: c93460ab4524c8b4492a8cef5629057b7428ecee