Option to skip CopyPropagation pass on methods with many registers
Summary:
The CopyPropagation pass uses non-linear data structures to represent information about registers. This may cause excessive memory usage. This change introduces a new configuration option with a default bound on the maximum number of registers in a method that the pass will operate on.

This comes at a small cost (slightly fewer instruction eliminated), but also dramatically speeds up the CopyPropagation pass.

Reviewed By: int3, justinjhendrick

Differential Revision: D14250920

fbshipit-source-id: 7f804487d4336a33f4b3dba6a57fe41cf9c3b513