Compile error on Mac in @mode/dbg in liblocator.
Summary:
Sometimes it's useful to be able to build an android app in mode/dbg, to get full symbols when debugging, on a Mac, where use of emulators is easier.  I've discovered that this doesn't work; there are things that don't compile in mode/dbg.

This diff is an error (at least according to the default build flags) in an assert.  The build command I was trying was:

cd ~/fbsource/fbandroid; buck build mode/dbg --deep mobilehome

That failed (after fixing the issue one down in this diff stack), with:

Command failed with exit code 1.
stderr: native/liblocator/locator.cpp:88:20: error: comparison of integers of different signs: 'int' and 'const uint32_t' (aka 'const unsigned int') [-Werror,-Wsign-compare]
  assert(pos - buf <= encoded_global_class_index_max);
         ~~~~~~~~~ ^  ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

Reviewed By: tophyr

Differential Revision: D14295776

fbshipit-source-id: 3c42ccc238efa1a3b9b2bca12d438b91695bf47f