Cold-start class splitting for static methods
Summary:
This is part one: Splitting out static methods with weight 0 in cold-start.
(To follow: Part two, splitting non-true virtuals, and part three, splitting true virtuals via a trampoline, adapting ideas from D7610256.)

The approach here is a new interdex plugin. This enables...
- only treating classes that end up in the primary / cold-start dexes;
- accounting for extra classes (and later methods when introducing trampolines), which is important to determine when a dex is full

Relocated methods are moved into new special classes at the end of the dex. Each class is filled with up to a configurable number of methods; only when a class is full, another one is created. Separate classes are created for distinct required api levels.

(This whole approach is very different from the stack of D13985945, where class splitting is done for non-cold-start dexes with the goal of reducing cross-dex-references.)
Integrating some changes from D13595508 and D13980133 to determine if an InterDex plug-in doesn't like it when methods of a class get relocated (namely, SwitchInlinePass).

Reviewed By: emma0303

Differential Revision: D14160853

fbshipit-source-id: 61a6f854f671a9b114090c5e084613187bbada4f