Treat INSTANCE_OF correctly
Summary:
As title.

suree404 found this bug. INSTANCE_OF dest reg holds a value of 0 or 1,
depending on the result of the check and we shouldn't associate it with any
type. This was an oversight of mine, when I added the analysis :(

The fix is to treat the destination register as on the default case.

Differential Revision: D14356325

fbshipit-source-id: 98c34ccfcd03fcbe75e3a5717c53163f2389e9aa