When determining which class to use to seed a dex, don't use a generated class.
Summary:
Background: The cross-dex-ref-minimization algorithm fills up each dex with classes that tend to share the same references. However, at the beginning of each new dex, there is a blank slate. At that point, the algorithm picks the class with the most references --- the "worst" class. This tends to give good guidance for filling the remainder of the dex (it increased the efficiency of the algorithm), and it may provide some build-over-build stability.

The change in this diff is: When determining which class to use to seed a dex, try to not pick a "generated" class.
"generated" classes mainly include those which are produced by type erasure; they tend to have a non-stable numeric name, and their structure may change in ways only indirectly to a developer's code change.

Advantages of this change:
- "generated" classes tend to be not stable and may cause drastic build-over-build changes; so removing them from considering for the worst seed class may make build sizes less volatile. (Admittedly, an unproved hypothesis.)
- Excluding generated classes with non-stable names makes enables a curated lists of seed in the future (to possibly further improve wins and/or reduce build volatility by hard-coding instead of finding at build-time what class to seed each dex with)
- It actually happens to be an (uncompressed) size win, so why not...

The same change is applied here to redex-stable and (-unstable).

Reviewed By: wsanville

Differential Revision: D14428787

fbshipit-source-id: 8c355a6a92d82008ff99f01b80ef1c926c3eea99