Avoid matching the same const-load insn to multiple annotated constant values
Summary:
There's a bug in the constant lifting logic.
If the MethodMeta annotation has two constant values sharing the same type and value, e.g., MethodMeta("II", "1:1"), we could match the same instruction to both of them. This will lead to a crash in `IRList` later on since we will try to replace the same instruction multiple times.

- I added a simple check when parsing all relevant const-loads. We don't match the same instruction more than once.
- Added a test case in the `instr:typeerasure_advanced` test.

Reviewed By: emma0303

Differential Revision: D14573873

fbshipit-source-id: 5d6b15c727bad3e013e97c7e921f8adc31cccb78