Support for the simplest version of Android Bundles (#390)
Summary:
As per https://github.com/facebook/redex/issues/367, we will need support for Android Bundles pretty soon and I was hoping we can get at least some rudimentary support into ReDex sooner rather than later.

The submitted patch works for us, but I am not sure if it's generic enough for the general audience. The missing bits was to tell the unpacker where the dex files are located in the archive, which is needed to (1) determine secondary dex mode, (2) unpackage dex files to a separate directory and (3) copy ReDex-processed dex files back to the directory containing unzipped archive.

I would be very happy to work on this further given some feedback on how to finesse/generalize it if need be.

I also made a decision of distinguishing the "mode" (APK or bundle) based on file extension - other options are clearly a possibility as well, though.
Pull Request resolved: https://github.com/facebook/redex/pull/390

Reviewed By: thezhangwei

Differential Revision: D14630061

Pulled By: Feng23

fbshipit-source-id: 7d391e5ecd508401beb56cd86fbfc8edede30f8b