Cross-dex-ref minimization: Include some skipped classes in frequency considerations.
Summary:
Skipping a class due to a plugin might mean that (members of) of the
class will get emitted later via the additional-class mechanism,
which is accounted for via the erased_classes reported through the
plugin's gather_refs callback. So we'll also sample those classes.

The result is an increased efficiency of the cross-dex ref minimization algorithm.

Reviewed By: int3

Differential Revision: D13806818

fbshipit-source-id: a1a5b37c40eebcd3db734f96a665f53324f8dbad