Don't relocate methods with no_optimizations() set, and respect api levels.
Summary:
This replaces the earlier idea of just not handling classes or methods which have annotations on them (which was somewhat futile as many annotations already got erased when the InterDex pass runs).
Also, put relocated methods into classes in a way that all methods in a class share the same api level.

Reviewed By: int3

Differential Revision: D13980133

fbshipit-source-id: 06ee7f53c37f839a948ac7e26b791f6028364eca