Ignore all relocation target classes for weight purposes
Summary:
While there's a good amount of methods that can be potentially relocated, most of those methods end up not actually getting relocated, as their affinity to their original host class is very high.

In this change, we reflect this by not including relocated target classes in the weight calculation that flows into the cross-dex-ref-minimization priority.

As a result, we get a bigger win from the method relocation feature.

There was already a provision to ignore the relocated target class in some cases; this is basically getting generalized:
- Before this change, a new relocation dummy class was not included in the weight computation of the particular relocated method that was moved into that relocation dummy class. However, all other classes referencing that method would still account for the weight of the relocation dummy classes of relocated methods.
- After this change, no relocation dummy classes get included in any weight accounting.

Reviewed By: int3

Differential Revision: D14759417

fbshipit-source-id: 8022c4a0885eaf47873ae28a3439ff7a6e3af625