Mitigate issues caused by dexlayout inflating debug information
Summary:
In Android 8+ there's a new system called dexlayout that's leveraged when PGO compilation on device occurs. This system inflates dex files into an IR so that it can move methods around. When creating IR dexlayout also inflates debug information. For each position entry that a debug program emits a 64bit model is created and then a pointer to that model is inserted into a vector. This means the runtime physical RAM cost of debug info with N entries is 12 * N bytes (we operate in 32 bit mode so a point is 4 bytes) and the VA cost is 2^ceil(log_2(N)) (this is due to vectors resizing by doubling which I confirmed to be the case in standard AOSP).

Unfortunately, it looks like theres a bug in ART where debug info that's shared across multiple methods gets reinflated. Thus, for M methods that share debug info with K entries there will be K * M position entries emitted, i.e. the N above is K * M.

In order to mitigate OOMs that we've observed we put an upper limit on the N = K * M in the above calculations for given debug info and emit multiple debug programs that fit closer to the size of the functions using those debug programs in order to shrink the K * M factor for each debug program. Right now each debug program is limited to 8KiB inflated position entries, with the total number of position entries limited to 2MiB for each arity.

Reviewed By: justinjhendrick

Differential Revision: D14692913

fbshipit-source-id: 3c01cdc7dd37b9f82cc8bbc6945fa979d604e0df