Fixed symbolicator app that was not reporting correct class/method names in case of multi-level inlining (#393)
Summary:
I am not sure if anyone uses `tools/debug-info/symbolicate-trace.cpp` in real life or it's more of a template to be modified by interested parties, and so I am not sure how important this fix is. Nevertheless, with multi-level inlining, the previous version will not symbolicate the stack trace correctly as it will re-use class name and method name from the leaf (inlined) method, even though ReDex generates all the required information into the map (V2).

As an aside, in presence of ProGuard inlining, even with this fix the stack will be partially incorrect as the source file information will not be recovered correctly. This can be done the same way that ProGuard does it, that is using class name to re-create the file name. I can put in another PR for that if there is interest.
Pull Request resolved: https://github.com/facebook/redex/pull/393

Reviewed By: Feng23

Differential Revision: D14816037

Pulled By: justinjhendrick

fbshipit-source-id: 84b095d74440a3fad8531402d3ada4f7ff9743a7