Change method spec in-place.
Summary:
Step 1: Collect all the field/method definitions and type/field/method references regarding to each candidate interface at the analyzing step.

Step 2: Escape the interfaces that we could not merge them into their single implementation classes (See the `EscapeReason`).

Step 3: Rename methods if they will collide to others after updating the interfaces to their implementation types.

Step 4: Do the optimization by updating all the definitions and references that we collected at analyzing step.

**Some implementation details**
When we update a method, we **create a new method** and move the code and all the attributions of the old one to the new one, and **maintain a map from old to new**.
With the map, we should also look up the map when do the update since an old method might have already been updated, we should do the update based on the previous updated one.
**To avoid looking up the map recursively, we never put a method into both `methoddefs` and `intf_methodrefs`. (That's probably the reason why we don't support removing an interface if it contains a method that references the interface itself.)**

Eliminate the old to new map, and only keep the map from interface method to implementation method since it's the only part we want to do a real "replacement".

Removing more interfaces.

Reviewed By: thezhangwei

Differential Revision: D14854753

fbshipit-source-id: dd8485771adff9f03e403ff089be00baca956158