Making remove-unused-args deterministic for non-virtuals
Summary:
There used to be a lock that I found a bit dubious (it was supposedly brought in to fix an issue, but the underlying parallelization always splits work at class boundaries, so I don't see what the lock would fix in practice).

Even if the lock would have done something, it would likely have introduced some form of non-determinism. Re-enabling handling of non-true virtuals exacerbated this non-determinism in the way it handled renaming.

This change removes the lock, and instead introduces an explicit sequentialization for renaming operations, while still performing everything else before and after in parallel.

Reviewed By: int3

Differential Revision: D14882775

fbshipit-source-id: 941c15b03b333d20459e01a98c09ce8d6b9f2d2a