Test data for manifest with exported attribute as a reference
Summary:
I'm going to bootcamp T39834130, but before doing so I'd like to supply the person who picks up the task with an example input for test cases.

As stated on https://github.com/facebook/redex/issues/368, we don't correctly parse the AndroidManifest.xml in all cases if a boolean literal isn't supplied for the `android:exported` attribute.

Reviewed By: newobj

Differential Revision: D14899052

fbshipit-source-id: a88829971e922b007015216e5bf372fcb6be166a