Back out "[redex][singleimpl] Remove NEXT_PASS"
Summary:
Original commit changeset: 93c359175110
Since we still have a map from old_method to new_method`m_intf_meth_to_impl_meth`, and we don't look up the map recursively. We still need put some interfaces to next round of optimization.
Example:
If we try to remove interfaces I1, I2 and I3, and we have an interface method `I1.method(I2, I3)`.
When we try to remove I2 or I3, we could not remove I1 in the same round.

Differential Revision: D14938702

fbshipit-source-id: 5c55c2a464cfdf9c0e6edcb1c9b262a137bd7098