update redex.py to read redex pg mapping output file name
Summary: There's an argument `class_rename` in the `RenameClassesPassV2` section in quite a few redex config files but `RenameClassesPassV2` never actually read that. Instead, the string of the output file name is hardcoded in the python script. This diff changes the script to read the config. If the `class_rename` argument is not there, default the path to the old one that's hardcoded.

Reviewed By: minjang

Differential Revision: D14956260

fbshipit-source-id: edf2700295f7bc0bb83dc4e19abc51353bb92451