Make local alias / escape analysis more generic
Summary:
The analysis used by OSDCE can be useful for a variety of other object-sensitive analyses. This diff allows us to plug in a generic implementation of a `Store` (heap) so that we can use a variety of abstract object models without repeating the alias/escape logic.

In the interest of keeping compile times down, I've opted to create an `EnvironmentWithStore` interface that type-erases the template `Store` type. This enables us to write methods that handle these Environments generically without including a lot of code in C++ header files. Local testing of OSDCE suggests that the virtual method call overhead is negligible.

Reviewed By: NTillmann

Differential Revision: D14908913

fbshipit-source-id: 83f9bb38cf0a521d2833e46cc0265c26a5ae81e5