Track additional DexTypes an AbstractObject is likely to be
Summary:
The analysis is currently unable to take advantage of the type reference from an instance-of instruction.
This leads to us losing the opportunity to report the precise type information on certain reflection operations.
Here I'm trying to attach a set of DexTypes to the AbstractObject keeping track of its potential DexTypes aside from the one we know for sure.
Currently we populate this information from `instance-of`s. It's up to the user of the analysis how to interpret the result.
It's simple since we don't have to only propagate this type information to the taken branch of with respect to the result of the `instance-of` check.

Reviewed By: Feng23

Differential Revision: D14943446

fbshipit-source-id: 8852d9107c53419a6d774be9af09440adc46b99f