Relocating type checking logic out from TypeInference
Summary:
The TypeInference inside IRTypeChecker operates in two modes. It analyzes the types on each instruction during the fixed point iteration. After reaching a fixed point, the type checker then uses it to perform type checks. The arrangement is mixing two purposes in one place and not so great for modularity if we want to reuse the type inference part.

In this diff, I remove the `m_inference` flat from TypeInference and pull out its the type checking logic. The change paves the way for us to move TypeInference to its own file and make it reusable in other places.

Reviewed By: NTillmann

Differential Revision: D14988736

fbshipit-source-id: aff08bc2ba7954047629e747ba70e83ea5379e5a