make marking all annotations as keep in ProguardMatcher a configurable behavior
Summary: Originally, ProguardMatcher will just mark all annotations as keep. This is causing the renamer unable to rename annotations at all. This should not be necessary because we can just whitelist any annotation we want from removal or renaming just like classes, methods and fields that are being reflected on.

Reviewed By: int3

Differential Revision: D15019030

fbshipit-source-id: fc7d548bc5acd7152009c9c32c7a597508b5cd8c