Add a blacklist to the Bridge pass
Summary:
See T43077386. The bridge pass ends up inlining a call to a private
method and the caller doesn't have access. The proper fix would be to
update the access flags at inlining time. MethodInline is doing that,
but it is using the CFG inliner. This pass is not using CFG yet, so
for the time being I am only adding a blacklist in Bridge pass, seems
like this could be handy anyway.

Reviewed By: thezhangwei

Differential Revision: D15021667

fbshipit-source-id: 88d04c41a39765bed2bdbac51f99c21b65b730b2