Removing clobbering limitation of up-code-motion pass
Summary:
When both targets of a conditional branch assign to the same registers, then the up-code-motion pass may move up the assignments of one branch before the if-instruction, if that allows to remove a goto.

A limitation used to be that the assigned registers must not overlap with any of the registers that feed into the if-instruction.

This limitation is here-by lifted, introducing new temp registers, extra move-instructions, and rewriting the if-instructions to refer to the temp registers.
In practice, almost all additional moves are eliminated by the register allocator / copy-propagation. So there's virtually no cost, but the benefit of the up-code-motion pass being much more effective.

Reviewed By: thezhangwei

Differential Revision: D15076490

fbshipit-source-id: 214d75168b3f6f9747b63d3e2829b3f4c50b1bc0