Sprinkle const on a bunch of DexItem APIs
Summary: This is technically not "correct" because the const values we pass in may be stored and later retrieved as non-const values, but that's a common problem with storing const types in containers, especially a global one. Having `const` APIs for things that don't actually mutate their parameters themselves allows many more things to be marked `const`, for a small increase in unsoundness during compilation.

Reviewed By: minjang

Differential Revision: D14909124

fbshipit-source-id: 631fe157a0b5f1d276164d823e5c3c3294ce931e