Wiring through architecture to enhance ReduceArrayLiterals pass
Summary:
Before KitKat, the Dalvik x86-atom backend had a bug in the implementation of the filled-new-array instruction which greatly reduced the impact of the new ReduceArrayLiterals pass for min_sdk < 19. TThis is now further getting gated by the architeture: We only need to hold back the optimization when we don't know the architecture, or if it's x86.

Added unit test for ReduceArrayLiterals pass to reflect this check.

Reviewed By: int3

Differential Revision: D15113466

fbshipit-source-id: 151532ff50f4979d4b69122f64f05c9a1e97640d