Account for Dalvik verifier bug for filled-new-array
Summary:
The Dalvik verifier had a bug for the case of arrays-of-arrays:

It retrieves the "element class" to check if the elements are of the right type:
https://android.googlesource.com/platform/dalvik/+/android-cts-4.4_r4/vm/analysis/CodeVerify.cpp#3191

But as this comment for aget-object indicates, this is wrong for multi-dimensional arrays:
https://android.googlesource.com/platform/dalvik/+/android-cts-4.4_r4/vm/analysis/CodeVerify.cpp#4577

The "fix" is to not emit filled-new-array instructions for arrays of arrays when targetting a min-sdk before Lollipop / Android 5 / API-level 21.

Fun fact: This is exactly the bug that I fixed in D15025972 in our own verifier/register allocator.

Differential Revision: D15204719

fbshipit-source-id: f34f2e14f569dbeaa1cfb84dd49476ca94784c03