Fix subtle bug in reduce-gotos pass
Summary:
The reduce gotos pass may shift all registers, and then it needs to also account for this in the liveness information that is computes.

The fix makes the computation of the liveness information lazy; maybe it's an overall speed up.

No known issue ever came out of this; I just found it by auditing code for other reasons.

Reviewed By: justinjhendrick

Differential Revision: D15245668

fbshipit-source-id: 16067b7dd8bffe0449e5cc88be07e3fa08bfa8e8