Bump API levels 14 & 15 (ICS) up to 16 (Jelly Bean)
Summary:
Google has removed ICS support from Play Services and from Android NDKs after r17.

In order to pave the way for NDK upgrades beyond r17, we are nudging all apps to bump minSdkVersion to 16.

This diff will land as soon as you accept it, via
If there is no response before May 15, we will go ahead and land it.

Reviewed By: justinjhendrick

Differential Revision: D15266449

fbshipit-source-id: ec735ad62f0d6848e0631ec6655acbb5fb6f712e