Making StaticReloV2 more deterministic
Summary:
The order of candidate matters for the order of the vertices, which in turn matter for im which order methods get relocated, which in turn might affect some internal renaming decisions.

I found that the StaticReloV2 pass is non-deterministic while working on a per-pass deep stable hashing feature (D14493342). Not sure if this is (the only) root cause of non-determinism in StaticReloV2, but it seems relevant.

Reviewed By: Feng23

Differential Revision: D15305226

fbshipit-source-id: 4d6add89f31821e34bdc0cd5616b3238a53f0864