Making AccessMarking less non-deterministic
Summary:
Compute an ordered representation of methods to privatize. This matters, as
the dmethods and vmethods are not necessarily sorted, but add_method does
a best-effort of inserting in an ordered matter.
But when dmethods and vmethods are not ordered to begin with, then the
order in which we attempt to add matters.

I found that the `add_method` in `mark_methods_private` of the AccessMarking pass is non-deterministic while working on a per-pass deep stable hashing feature (D14493342). Not sure if this is (the only) root cause of non-determinism in AccessMarking, but it seems relevant.

Reviewed By: emma0303

Differential Revision: D15328481

fbshipit-source-id: 45480de0249589c6ae283edcbdccb54bae68c750