Improve bridge pattern matching
Summary:
The pass's pattern matching is simple, and we may get some false-positive
bridge-bridgee pairs. It's not a big deal, since it's just a special kind of
inlining.
The problem is that it may find bridgee-bridgee methods from different classes
and then inline the bridgee to bridge without proper access changing.

The diff keeps the simple reasonable inlining logic but adds additional checking
on whether the bridge and bridgee are in the same class.

Reviewed By: emma0303

Differential Revision: D15329107

fbshipit-source-id: ac966a79b5e4d97da2643bd817c33dd5bdf33032