In CSE, don't include "safe" methods as barriers
Summary:
This diff lists a number of framework methods which are...
 - static, or
 - direct (constructors), or
 - virtual methods defined in final classes

and that do not mutate any fields or array elements that could be directly
accessed (read or written) by user code, and they will not invoke user code.

The list of methods is not exhaustive; it was derived by observing the most
common barriers encountered in real-life code, and then studying their spec
to check whether they are "safe" in the context of CSE memory barriers.

(Some items are commented out; that's because they are virtual methods in
non-final classes. They could be enabled once we can track the exact
receiver type. This would double the win.)

Reviewed By: justinjhendrick

Differential Revision: D15263567

fbshipit-source-id: 4414231635cc5307871336332c1187cba89c049e