Making ReduceArrayLiterals less non-deterministic
Summary:
The order in which array literals is being processed does matter for determinism, as the allocation of registers depends on the order.

Also "upgraded" some iterators to be cfg-level iterators, for smaller code size.

I found that the ReduceArrayLiterals pass is non-deterministic while working on a per-pass deep stable hashing feature (D14493342). Not sure if this is (the only) root cause of non-determinism in ReduceArrayLiterals, but it seems relevant.

Reviewed By: wsanville

Differential Revision: D15325529

fbshipit-source-id: cc96e71326c4b37b0a7695354c5b958575f94b60