DexHasher to compute stable hashes of scope after each pass
Summary:
This helps finding non-determinism.
What is getting hashed is... Everything that defines the scope! In order. This is much more insightful to detect subtle non-determinism, and identify the pass that introduced it.

Boost hashes are twice as fast as SHA1, and good enough for our purposes.
Hashes are outputted into the redex stats in their own section, including an initial hash, as well as into each pass' section. Look for the ~result~hash~ hash (it's a 52-bit number to make JSON happy) that represents the final state the scope after the pass.

Look for stats-bot results in the stack diff to see an example of how it will look like.

The time overhead of hashing is reasonably small, so I want to enable it by default.

Reviewed By: int3

Differential Revision: D15337362

fbshipit-source-id: 77cd04b5f642f024bedd86f8732ab70e46416a95