Update the upstream CMake googletest dependency.
Summary:
SPARTA's gtest was configured to use gtest's 'master' branch from the
googletest repository.  Currently SPARTA requires the 1.8 version of gtest.
This patch pins the version so that a configure of SPARTA will obtain the
desired version of gtest (1.8.1).  This patch avoids deprecation warnings when
building the SPARTA with the newer gtest's master sources.

Reviewed By: int3

Differential Revision: D15387075

fbshipit-source-id: 76ccb74fb6c222c7bdb528ea25ad6b7c977b1277