convert AnnoKillPass to Configurable
Summary:
Convert AnnoKillPass to Configurable. This shows a couple things:

1) The binding definitions in AnnoKillPass are correctly inherited by SystemAnnoKillPass
2) Add support for a very complicated primitive (map of vector of strings)

Also add AnnoKill stuff to dextricks app to demonstrate it working end to end

I filed T44241164 to track that the type enforcement for this type only comes out as "dict" on the Python side. Redex the binary will complain more specifically, but I still think it'd be good to have richer type verification in the Pyhton API.

NOTE: in later diffs, I want AnnoKillPass to bind directly to a map of DexType -> vector of DexTypes, as this is the final form AnnoKill uses for its data. Baby steps first though.

Reviewed By: justinjhendrick

Differential Revision: D15299281

fbshipit-source-id: 99f0097ce5eb4dc9e21cb1beaa69b9e35efac089