Fix division for address offset in debug info op
Summary:
While using the dex library I was getting back floats as addressses.
The result of division in Python is a float, but we want the floored int.
Casting it should help. I checked the Python code for other divisions and noted
the `int()` cast is used there as well.

Reviewed By: int3

Differential Revision: D15418691

fbshipit-source-id: 050e3108f5106f08b2ca4153ce1dc9c67fb4bc8f