remove config option 'stats_output'
Summary:
Motivation:

Redex generates many output files. Typically named with something defined in the redex config. However, this is not really convenient because we assume files names to be default in a lot of places. redex.py also does this and copies files while renaming it with a `redex-` prefix. We should probably better off be using the desired name as default.

Since we should be consistent with our naming, i.e. the redex generated artifacts should start with `redex-` (currently 'iodi_metadata' is not, but we can make a exception for compatibility) , I don't see a reason why we can't generate these names in the first place.

Reviewed By: int3

Differential Revision: D15400783

fbshipit-source-id: 418859af194bc92241150522bfbabc94ab3ec314