Enum values may not escape through some invocations
Summary:
Introduce a `ParamSummary` structure for a method which indicates how the parameters are used in the method body.
If a candidate enum object is passed as an `Object`(super types) parameter of a method invocation, we believe there is an implicit upcating before.
But with the param summary information of the method, we may know that the parameter is not escaping from the method or only escaping as return value, the invocation is not black box any more and we can do further analysis with the information.

Example:
```
// The parameters are only escaping as return value or not escaping. We store the indices of params to `safe_params` set.
// The return value is always the first parameter.
// So the param summary of the method is "safe_params: [0,1], returned_param: 0".
public static <T> T notEscape(T obj, String s) { return obj; }

public void method(CandidateEnum obj) {
   // Now we know more about the invocation
  CandidateEnum obj2 = notEscape(obj);
}
```

Apply escaping analysis to get the `ParamSummary` of some static methods.

Reviewed By: int3

Differential Revision: D15402147

fbshipit-source-id: 98580e5e08e807e31030825b4f83e7ce535fee81