Bindings for GlobalConfig and its subconfigs
Summary: Create classes and bindings for GlobalConfig and its constituent parts. The actual /use/ of the global config (via ad-hoc JsonWrapper lookups in the code) is not altered yet. The point of this diff is to cause the Python API to be generated for the global config, which would allow total configuration of any app by refig.

Reviewed By: yuxuanchen1997

Differential Revision: D15520266

fbshipit-source-id: 6b72395a0ebca856438298514216d5eb0a87fb91