Track different locations separately during abstract interpretation
Summary:
Before, we only used method-level information about locations that are read, and a value was either "barrier sensitive" or not.
This is now getting refined:
- several bits of each value id are now used to represent the different locations from which was read to construct this value.
- when encountering a barrier, only those values get invalidated which are affected by that particular barrier.

This increases the precision of the analysis, and allows for some heap-dependent values to selective survive a barrier.

Also, the CommonSubexpressionElimination class with all its nested structure was getting too unwieldy. This is getting refactored and all the details are put behind a `cse_impl` namespace.

Reviewed By: int3

Differential Revision: D15388776

fbshipit-source-id: 483892e5ab5d45b56c99967e14b291670fdc9e54