Store-to-Load forwarding
Summary:
After a (non-volatile) [ais]put, we record the fact that the stored value is now equivalent to an [ais]get.
This in effect realizes general store-to-load forwarding. As one can see in the redex stats bot result, this makes the more targetted version of this optimization in the peephole pass ineffective.

Reviewed By: Feng23

Differential Revision: D15388974

fbshipit-source-id: 09ed8d6475416365bd1cab7e4bc11b605565831a