Speeding up CSE by enhancing patricia tree functionality
Summary:
The most expensive thing in the Redex CSE implementation is the loop over all known values.
It checked for each key if a bit is set, and then later erases such elements one by one.
This is all a bit silly given the underlying patricia tree representation

The change here is to expose a new way of bulk-erasing all elements where keys match a given bit mask, and using that in CSE.

This brings the complexity of that CSE loop from O(n*log n) down to sublinear.

Reviewed By: int3

Differential Revision: D15559929

fbshipit-source-id: 83ef0c47ed735143cd918bd1d0aa85cb3d469a3e