Merging of virtual methods
Summary:
VirtualMergingPass removes virtual methods that override other virtual methods, by merging them, under certain conditions.
- we omit virtual scopes that are involved in invoke-supers (this could be made less conservative)
- we omit virtual methods that might be involved in unresolved invoke-virtuals.(not actually observed in practice)
- of, course the usual `can_rename` and not `root` conditions
- the overriding method must be inlinable into the overridden method (using standard inliner functionality)

When overriding an *abstract* method, the body of the overriding method is essentially just moved into the formerly abstract method, with a preceeding cast-class instruction to make the type checker happy.

When overriding a non-abstract method, we first insert a prologue like the following into the overridden method:

```
  instance-of               param0, DeclaringTypeOfOverridingMethod
  move-result-pseudo        if_temp
  if-nez                    if_temp, new_code
  ... (old body)

new_code:
  cast-class                param0, DeclaringTypeOfOverridingMethod
  move-result-pseudo-object temp
  invoke-virtual            temp, param1, ..., paramN, OverridingMethod
  move-result               result_temp
  return                    result_temp

...
```
And then we inline the invoke-virtual instruction, and remove the overriding methood. Details vary depending on the whether the method actually has a result, and if so, what kind it is.

For now, we only do this transformation within each dex (when overridden and overriding method are within the same dex), as it tends to shuffle around many dependencies across dexes otherwise, with not always positive effects.

Future work: Integrate with InterDex pass, maybe via plugin, to allow stuffing dexes further.

Reviewed By: thezhangwei

Differential Revision: D15466342

fbshipit-source-id: 95b179f793bd1f83d9ee7e23a953743e431999c7