Fix FinalInline's mis-encoding of static field values
Summary:
FinalInlineV2 was taking the values of the fields at <clinit> exit and making
them the encoded values of those fields. Then it would re-run constant
propagation on the <clinit> which would typically discover that the sputs that
wrote to those fields were now redundant, since they were writing the values
identical to the encoded ones.

The problem was that sometimes the clinit would read the existing encoded
values before writing a new value into the field. Not accounting for this meant
that the field read would now obtain a different value after FinalInline ran.

The "full" fix for this would be to avoid encoding values for fields whose
sgets are not all dominated by sputs. But since it's rare to have fields read
in their own class' static initializers, I'm just blacklisting all such fields.

Differential Revision: D15590869

fbshipit-source-id: 4225c84c1dea3a639f4ade341b62c96edf8d4aa9