Do not merge blocks when that would lead to verification errors
Summary:
This is a bug fix. The new test case illustrates the problem: An incoming register could be either a non-zero integer or a reference that gets checked with a if-nez instruction in two equivalent blocks. The blocks look dedupable, but the verifier will refuse to operate on a joined non-zero integer and reference.

The added overhead does not seem to make DedupBlocks pass run significantly slower overall.

Reviewed By: justinjhendrick

Differential Revision: D15684329

fbshipit-source-id: 971d25e0a19eedccc12695cc42c4ceeda2dca2c2