Speeding up CSE by enhancing patricia tree functionality, part 2
Summary:
(After the diff at the bottom of the stack) the most expensive thing in the Redex CSE implementation is the loop over all registers.
It checked for each register if a bit in the value is set, and then later erases such elements one by one.
As erasing an element is O(log n), the operation over all registers is O(n*log n), potentially creating a good number of unnecessary intermediate nodes in the Patricia tree while erasing elements.

The change here is to expose a new way of bulk-mapping all values, and using that in CSE.

This brings the complexity of that CSE loop from O(n*log n) down to linear.

Reviewed By: int3

Differential Revision: D15568136

fbshipit-source-id: 52fb4418229fc1885976099e7f8cc26c266aebe8