Split RebindRefs into ResolveRefsPass + RebindRefsPass
Summary:
First step into reorganizing method/field rebinding.

I'm separating the actual reference resolution logic away from the existing RebindRefs pass. The new pass is now called ResolveRefsPass. We run that early on making sure the rest of our passes don't have to deal with the distinction between a ref and a def. The rest of RebindRefs focuses on coalescing method references to a smaller set, and it takes place right before InterDex.

There's a series of improvements we can add to both of the two passes. Ref resolution can be more precise if we incorporate more type information from analyses like type inference and point-to analysis. That helps to build a more accurate call-graph later on. Ref rebinding can be more aggressive or cover more targets than what we currently do.

More to come.

Differential Revision: D15608700

fbshipit-source-id: abba0603457d7d05a4b5b8923089718f2e19dcdf