Make lookup table analysis compatible with external method ref binding
Summary:
The lookup table analysis currently check against the method reference of the invoke-virtual instruction on the `ordinal()` method.
However, with the ongoing work on method reference binding, the method ref can be resolved to `java.lang.Enum.ordinal()` at an earlier pass. This change could break the analysis. So we need to make lookup table analysis more flexible or ideally more sophisticated.

Here is a simple approach where we add special checks on the JDK `Enum` type. Alternatively we could beef up the `EnumInSwitch` analysis to incorporate type inference. Which will be a significant larger change. Perhaps that could be a long term goal.

Reviewed By: Feng23

Differential Revision: D15954907

fbshipit-source-id: dc583e7119d2ecbfb4642dadc8d054367f02afa7