Resolve method ref to external def
Summary:
Add options in ResolveRefs and RebindRefs to allow rebinding to an external method reference.
We ignore these cases before to avoid dealing with the discrepancy between different external dependencies on various OS versions. But we also gave up on the opportunities for saving method ref count.

Simplify allowing rebind to an external method def obviously doesn't work. The external `android.jar` we are compiling against maybe not reflect the actual API available on older OS versions. Therefore we have to be very careful about rebinding reference to Android runtime classes. Specially when we are dealing with the logic in Android support library or Androidx. I've experimented with a few tweaks. What appears to work for me is to avoid rebinding of a reference pointing to external classes. This is what we use in this diff. Note that this is different from allowing rebind to an external def, which is the point of this diff. The new logic is gated by a config in the two passes respectively. We can turn them off if we see an unexpected error.

Reviewed By: Feng23

Differential Revision: D15929078

fbshipit-source-id: 01c370f961f1f2f3d274ffdf4f106d51e9f1b88b