unit tests: Add EXPECT_CODE_EQ macro
Summary:
We compare IRCode objects by serializing them first. However, the serialized
forms lack newlines between instructions and so are rather difficult to read.
It's nice to print the original IRCode objects which have those newlines. Folks
have been doing this printing on an ad-hoc basis in a couple of places. This
diff adds a macro so it's convenient and terser.

Unfortunately this doesn't handle the pattern where a test serializes an IRCode
before running some transformation on it, to check that the transformation
hasn't changed the input. Could certainly handle this via some overloading but
I'm lazy to do it right now.

Reviewed By: Feng23

Differential Revision: D16064698

fbshipit-source-id: 75cd41c4ebdb5fb89000783de961f6927d61977c