Don't let instruction dest overlap wide src registers
Summary:
This causes a bug on some OS versions: we've observed it on the ART interpreter
as well as on Android 4.x. Previously, we added support for avoiding this
overlap when both the src and dest registers were wide values, but it appears
that the issue can also occur when the dest register isn't wide. This diff
addresses that.

Reviewed By: Feng23

Differential Revision: D16064699

fbshipit-source-id: 5cdad0bf119f1a18a60f42868bf5e971bbc4dbdd