Constant Propagation for Switch
Summary:
This patch tries to optimize a switch statement if the switch value is evaluated as a constant. While collecting its successor blocks from the switch, dead labels are removed in place. As a result, the switch can be optimized in two ways:
	- Eliminates the switch when only the default block is reachable.
	- Replaces the switch by a goto statement when only a non-default block is reachable.
The subsequent dead-code elimination cleans up the unreachable blocks that no longer have a label or an edge.

Reviewed By: NTillmann

Differential Revision: D16019671

fbshipit-source-id: 28cd81230076a5b92483c970903a2b93ac7c147d