Avoid assert in CSE due to referenced but not defined safe method
Summary: These asserts were introduced in D15395791 and triggers if redex is running without full framework available. If framework availability is a requirement, it should be checked before CSE and report error instead of assert in CSE. But from what I see in the code it is safe to mark referenced but not defined safe methods as "safe" in CSE because even if they are present, their bodies are not checked.

Reviewed By: NTillmann

Differential Revision: D15983445

fbshipit-source-id: ab398bf9b75e86cf55d9414c91a5e80f6f05019d